content/html/content/test/test_bug566064.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=566064
-->
<head>
  <title>Test for Bug 566064</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=566064">Mozilla Bug 566064</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 566064 **/

// This test is only checking the IDL/content attribute of 'formtarget'.
// The behavior is tested in test_bug566046.html.

function isFormTargetEquals(aElement, aValue, aShouldBeNull)
{
  if (aShouldBeNull) {
    contentAtttributeValue = null;
  } else {
    contentAtttributeValue = aValue;
  }

  is(aElement.formTarget, aValue,
     "formTarget IDL attribute value should be " + aValue);
  is(aElement.getAttribute('formtarget'), contentAtttributeValue,
     "formTarget content attribute value should be " + contentAtttributeValue);
}

function checkFormTarget(aElement)
{
  isFormTargetEquals(aElement, "", true);

  aElement.formTarget = "foo";
  isFormTargetEquals(aElement, "foo");

  aElement.setAttribute("formtarget", "bar");
  isFormTargetEquals(aElement, "bar");

  aElement.removeAttribute("formtarget");
  isFormTargetEquals(aElement, "", true);
}

var button = document.createElement('button');

ok('formTarget' in button, "formTarget is a HTMLButtonElement property");

checkFormTarget(button);

</script>
</pre>
</body>
</html>
